+++ mobile/android/android-components/components/concept/fetch/src/main/java/mozilla/components/concept/fetch/Headers.kt
@@ -65,6 +65,7 @@ interface Headers : Iterable<Header> {
      */
     object Values {
         const val CONTENT_TYPE_FORM_URLENCODED = "application/x-www-form-urlencoded"
+        const val CONTENT_TYPE_APPLICATION_JSON = "application/json"
     }
 }